+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # CHANGELOG
 
+## Head
+* small bugfixes
+* cleanup
+
 ## 1.6.1
 * bugfix (now correctly checking `theme_rename`)
-
-## 1.6
 * add error handler that asks the user to continue if an error occured
 * add google-sitemap-generator to preset plugins
 * remove dublicated plugin entry